requirements.txt
pandas==2.3.3
python-dotenv==1.2.1
SQLAlchemy==2.0.45
